Moving postgres to redshift and updating postgres to be like mysql and sqlserver and work.
