+++ postgres/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "leo-connector-postgres",
-	"version": "1.0.6-alpha3",
+	"version": "1.1.0-alpha1",
 	"description": "",
 	"main": "index.js",
 	"scripts": {
@@ -12,13 +12,14 @@
 	"dependencies": {
 		"aws-sdk": "^2.213.1",
 		"fast-csv": "^2.4.1",
-		"leo-connector-common": "^1.0.14",
-		"leo-sdk": "^1.0.66",
+		"leo-connector-common": "^1.0.15",
+		"leo-sdk": "^1.0.70",
 		"pg": "^7.4.1",
 		"pg-copy-streams": "^1.2.0",
 		"pg-format": "^1.0.4",
 		"pg-logical-replication": "^1.0.3",
-		"pg-native": "^2.2.0"
+		"pg-native": "^2.2.0",
+		"readable-stream": "^2.3.6"
 	},
 	"devDependencies": {
 		"chai": "^4.1.2",
