+++ sqlserver/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "leo-connector-sqlserver",
-	"version": "1.0.9-alpha3",
+	"version": "1.1.0-alpha1",
 	"description": "",
 	"main": "index.js",
 	"scripts": {
